2.1

+++ src/main/java/com/mycompany/librarymanagement/repository/UserRepository.java
@@ -14,8 +14,9 @@ import org.springframework.stereotype.Repository;
  * @author ASUS
  */
 @Repository
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, Integer> {
 
     public User findByEmailAndPassword(String email, String password);
+
 
 }
